(unfinished) cacheConfig() and updConfig() finish, FileUtil created and externalfile util folder refactored
