+++ src/main/java/io/github/codeutilities/commands/image/ImageParticleCommand.java
@@ -9,7 +9,7 @@ import io.github.codeutilities.images.ParticleImage;
 import io.github.codeutilities.util.ItemUtil;
 import io.github.codeutilities.util.chat.ChatType;
 import io.github.codeutilities.util.chat.ChatUtil;
-import io.github.codeutilities.util.externalfile.ExternalFile;
+import io.github.codeutilities.util.file.ExternalFile;
 import io.github.codeutilities.util.render.ToasterUtil;
 import io.github.codeutilities.util.templates.TemplateUtils;
 import io.github.cottonmc.clientcommands.CottonClientCommandSource;
