+++ src/main/java/io/github/codeutilities/util/file/FileUtil.java
@@ -1,0 +1,15 @@
+package io.github.codeutilities.util.file;
+
+import java.io.IOException;
+import java.nio.charset.Charset;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+
+public class FileUtil {
+    public static String readFile(String path, Charset encoding)
+            throws IOException
+    {
+        byte[] encoded = Files.readAllBytes(Paths.get(path));
+        return new String(encoded, encoding);
+    }
+}
